refactor(Sliding): 뮌, 비숍, 룩의 이동 추상화
